+++ alibi_detect/cd/preprocess.py
@@ -1,93 +1,94 @@
-# TODO: incremental PCA for online methods
 import logging
 import numpy as np
 from sklearn.decomposition import PCA
 import tensorflow as tf
 from tensorflow.keras.layers import Dense, Flatten, Input, InputLayer
 from tensorflow.keras.models import Model
-from alibi_detect.models.autoencoder import EncoderAE
-from alibi_detect.utils.prediction import predict_batch
+from typing import Dict, Union
+from alibi_detect.utils.prediction import predict_batch, predict_batch_transformer
 
 logger = logging.getLogger(__name__)
 
 
-def uae(X: np.ndarray,
-        encoder_net: tf.keras.Sequential = None,
-        enc_dim: int = None,
-        batch_size: int = int(1e10)) -> np.ndarray:
+class UAE(tf.keras.Model):
+    def __init__(self,
+                 encoder_net: tf.keras.Sequential = None,
+                 input_layer: Union[tf.keras.layers.Layer, tf.keras.Model] = None,
+                 shape: tuple = None,
+                 enc_dim: int = None
+                 ) -> None:
+        super(UAE, self).__init__()
+        is_tf_seq = isinstance(encoder_net, tf.keras.Sequential)
+        is_enc_dim = isinstance(enc_dim, int)
+        if is_tf_seq:
+            self.encoder = encoder_net
+        elif not is_tf_seq and is_enc_dim:  # set default encoder
+            input_layer = InputLayer(input_shape=shape) if input_layer is None else input_layer
+            input_dim = np.prod(shape)
+            step_dim = int((input_dim - enc_dim) / 3)
+            self.encoder = tf.keras.Sequential(
+                [
+                    input_layer,
+                    Flatten(),
+                    Dense(enc_dim + 2 * step_dim, activation=tf.nn.relu),
+                    Dense(enc_dim + step_dim, activation=tf.nn.relu),
+                    Dense(enc_dim, activation=None)
+                ]
+            )
+        elif not is_tf_seq and not is_enc_dim:
+            raise ValueError('Need to provide either `enc_dim` or a tf.keras.Sequential `encoder_net`.')
+
+    def call(self, x: Union[np.ndarray, tf.Tensor, Dict[str, tf.Tensor]]) -> tf.Tensor:
+        return self.encoder(x)
+
+
+class HiddenOutput(tf.keras.Model):
+    def __init__(self,
+                 model: tf.keras.Model = None,
+                 layer: int = -1,
+                 input_shape: tuple = None
+                 ) -> None:
+        super(HiddenOutput, self).__init__()
+        if input_shape and not model.inputs:
+            inputs = Input(shape=input_shape)
+            model.call(inputs)
+        else:
+            inputs = model.inputs
+        self.model = Model(inputs=inputs, outputs=model.layers[layer].output)
+
+    def call(self, x: Union[np.ndarray, tf.Tensor]) -> tf.Tensor:
+        return self.model(x)
+
+
+def preprocess_drift(X: np.ndarray,
+                     model: tf.keras.Model = None,
+                     tokenizer=None,
+                     max_len: int = None,
+                     batch_size: int = int(1e10)) -> np.ndarray:
     """
-    Dimensionality reduction with an untrained autoencoder.
-
-    Parameters
-    ----------
-    X
-        Batch of instances.
-    encoder_net
-        Encoder network as a tf.keras.Sequential model.
-    enc_dim
-        Alternatively, only the dimension of the encoding can be provided and
-        a default network with 2 hidden layers is constructed.
-    batch_size
-        Batch size used when making predictions with the autoencoder.
-
-    Returns
-    -------
-    Encoded batch of instances.
-    """
-    is_tf_seq = isinstance(encoder_net, tf.keras.Sequential)
-    is_enc_dim = isinstance(enc_dim, int)
-    if not is_tf_seq and is_enc_dim:  # set default encoder
-        input_dim = np.prod(X.shape[1:])
-        step_dim = int((input_dim - enc_dim) / 3)
-        encoder_net = tf.keras.Sequential(
-            [
-                InputLayer(input_shape=X.shape[1:]),
-                Flatten(),
-                Dense(enc_dim + 2 * step_dim, activation=tf.nn.relu),
-                Dense(enc_dim + step_dim, activation=tf.nn.relu),
-                Dense(enc_dim, activation=None)
-            ]
-        )
-    elif not is_tf_seq and not is_enc_dim:
-        raise ValueError('Need to provide either `enc_dim` or a tf.keras.Sequential `encoder_net`.')
-    enc = EncoderAE(encoder_net)
-    X_enc = predict_batch(enc, X, batch_size=batch_size)
-    return X_enc
-
-
-def hidden_output(X: np.ndarray,
-                  model: tf.keras.Model = None,
-                  layer: int = -1,
-                  input_shape: tuple = None,
-                  batch_size: int = int(1e10)) -> np.ndarray:
-    """
-    Return hidden layer output from a model on a batch of instances.
+    Prediction function used for preprocessing step of drift detector.
 
     Parameters
     ----------
     X
         Batch of instances.
     model
-        tf.keras.Model.
-    layer
-        Hidden layer of model to use as output. The default of -1 would refer to the softmax layer.
-    input_shape
-        Optional input layer shape.
+        Model used for preprocessing.
+    tokenizer
+        Optional tokenizer for text drift.
+    max_len
+        Optional max token length for text drift.
     batch_size
-        Batch size used for the model predictions.
+        Batch size.
 
     Returns
     -------
-    Model predictions using the specified hidden layer as output layer.
+    Numpy array with predictions.
     """
-    if input_shape and not model.inputs:
-        inputs = Input(shape=input_shape)
-        model.call(inputs)
+    if tokenizer is None:
+        return predict_batch(model, X, batch_size=batch_size)
     else:
-        inputs = model.inputs
-    hidden_model = Model(inputs=inputs, outputs=model.layers[layer].output)
-    X_hidden = predict_batch(hidden_model, X, batch_size=batch_size)
-    return X_hidden
+        return predict_batch_transformer(model, tokenizer, X, max_len, batch_size=batch_size)
 
 
 def pca(X: np.ndarray, n_components: int = 2, svd_solver: str = 'auto') -> np.ndarray: